+++ trecs/matrix_ops.py
