+++ trecs/tests/test_matrix_ops.py
@@ -1,6 +1,6 @@
 import numpy as np
 import test_helpers
-from rec.matrix_ops import normalize_matrix, contains_row, slerp
+from trecs.matrix_ops import normalize_matrix, contains_row, slerp
 
 
 class TestMatrixOps:
